feat(odyssey-react-mui): abstract preview into it's own component
